add all functions to ./lib/cpf.js
